Extract HomePage into a seperaate file
