+++ web/manifest.json
@@ -1,5 +1,5 @@
 {
-    "name": "Todo Later Web App",
+    "name": "Todo Later",
     "short_name": "Todo Later",
     "start_url": ".",
     "display": "standalone",
